iot2050-module-firmware-update: Add module firmware update
IOT2050 Customized external signal module, such as SM SENS DI. This kind
of module is not PLC1200 standard, but "customized" by IOT2050. Hence
such module could not be upgraded within the PLC ecosystem, but we have
to provide method to support the FW upgrading for them.

And the firmware update for the module should not be included with the
image. This is because the image becomes too large and cumbersome to
release when only the update tool needs to be released. Instead, the
module firmware update should be packaged separately in its own
Debian package for a more efficient and manageable release process.

Signed-off-by: Li Hua Qian <huaqian.li@siemens.com>